+++ clients/include/hipsolver_dispatcher.hpp
@@ -11,6 +11,7 @@
 #include "testing_geqrf.hpp"
 #include "testing_getrf.hpp"
 #include "testing_getrf_npvt.hpp"
+#include "testing_getrs.hpp"
 #include "testing_potrf.hpp"
 
 struct str_less
@@ -34,6 +35,7 @@ class hipsolver_dispatcher
         static const func_map map = {
             {"geqrf", testing_geqrf<false, false, false, T>},
             {"getrf", testing_getrf<false, false, false, T>},
+            {"getrs", testing_getrs<false, false, false, T>},
             {"potrf", testing_potrf<false, false, false, T>},
         };